# swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
  Cluster:
    properties:
      kubeconfig:
        description: Kubeconfig is an admin kubeconfig file for communicating with
          the cluster.
        type: string
        x-go-name: Kubeconfig
      name:
        description: Name of the cluster.
        type: string
        x-go-name: Name
      namespace:
        description: |-
          Namespace of the cluster. Models the Kubernetes concept of Namespace as OpenShift Hive
          uses that to allow multiple clusters with the same name, separated by owner.
        type: string
        x-go-name: Namespace
    required:
    - kubeconfig
    - name
    - namespace
    title: Cluster Cluster is a representation of a Cluster we will reconcile SyncSets
      to.
    type: object
    x-go-package: github.com/dgoodwin/syncsets/models
  cluster:
    properties:
      kubeconfig:
        description: Kubeconfig is an admin kubeconfig file for communicating with
          the cluster.
        type: string
        x-go-name: Kubeconfig
      name:
        description: Name of the cluster.
        type: string
        x-go-name: Name
      namespace:
        description: |-
          Namespace of the cluster. Models the Kubernetes concept of Namespace as OpenShift Hive
          uses that to allow multiple clusters with the same name, separated by owner.
        type: string
        x-go-name: Namespace
    required:
    - name
    - namespace
    - kubeconfig
    title: Cluster is a representation of a Cluster we will reconcile SyncSets to.
    type: object
    x-go-name: Cluster
    x-go-package: github.com/dgoodwin/syncsets/api
host: localhost
info:
  contact:
    email: dgoodwin@redhat.com
    name: Devan Goodwin
  description: |-
    Standalone Hive SyncSets without reliance on Kube API, CRs, and etcd.
    Schemes:
    http
    https
    Schemes:
    http
  license:
    name: Apache2
  title: SyncSets API.
  version: 0.0.1
paths:
  /clusters:
    post:
      description: TODO Some test description that should be expanded on someday.
      operationId: createCluster
      responses:
        "200":
          $ref: '#/responses/OK'
        "400":
          $ref: '#/responses/BadRequest'
      schemes:
      - http
      summary: Create a Cluster resource.
produces:
- application/json
responses:
  BadRequest:
    description: |-
      BadRequest indicates that there was an error in
      the HTTP request.
    headers:
      ResponseCode:
        format: int64
        type: integer
  OK:
    description: OK indicates that the HTTP request was successful.
    headers:
      ResponseCode:
        format: int64
        type: integer
  createClusterBadRequest:
    description: |-
      CreateClusterBadRequest BadRequest indicates that there was an error in
      the HTTP request.
    headers:
      ResponseCode:
        format: int64
        type: integer
  createClusterOK:
    description: CreateClusterOK OK indicates that the HTTP request was successful.
    headers:
      ResponseCode:
        format: int64
        type: integer
schemes:
- http
swagger: "2.0"